XWIKI-18972: Allow Skins to be loaded from the classloader
